added some new class testing files for unit tests, moved the unit test for gen random Board to dice controller test file
