+++ app/build.gradle
@@ -17,13 +17,26 @@ android {
             proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
         }
     }
+    testOptions {
+        // By default, unit tests throw an exception any time the code you are testing tries to access
+        // Android platform APIs (unless you mock Android dependencies yourself or with a testing
+        // framework like Mockito). However, you can enable the following property so that the test
+        // returns either null or zero when accessing platform APIs, rather than throwing an exception.
+        returnDefaultValues true
+    }
+
 }
 
 dependencies {
     compile fileTree(dir: 'libs', include: ['*.jar'])
+
+    // Required for instrumented tests
     androidTestCompile('com.android.support.test.espresso:espresso-core:2.2.2', {
         exclude group: 'com.android.support', module: 'support-annotations'
     })
     compile 'com.android.support:appcompat-v7:24.1.1'
+
+    // Required for local unit tests (JUnit 4 framework)
     testCompile 'junit:junit:4.12'
+
 }
