+++ app/src/test/java/com/example/swathi/gameboggle/BoardTest.java
@@ -1,0 +1,23 @@
+package com.example.swathi.gameboggle;
+
+import org.junit.After;
+import org.junit.Before;
+import org.junit.Test;
+import static org.junit.Assert.*;
+
+
+
+public class BoardTest {
+    @Before
+    public void setUp() throws Exception {
+
+    }
+
+    @After
+    public void tearDown() throws Exception {
+
+    }
+
+
+
+}
