+++ app/src/test/java/com/example/swathi/gameboggle/ExampleUnitTest.java
@@ -2,8 +2,6 @@ package com.example.swathi.gameboggle;
 
 import org.junit.Test;
 
-import java.util.ArrayList;
-
 import static org.junit.Assert.*;
 
 /**
@@ -15,20 +13,5 @@ public class ExampleUnitTest {
     @Test
     public void addition_isCorrect() throws Exception {
         assertEquals(4, 2 + 2);
-    }
-
-    @Test
-    public void random_board_generation() {
-        DiceController x = new DiceController();
-        ArrayList<String> y = null;
-        for(int i = 0; i < 5; ++i){
-            y = x.genBoard();
-            for (String s : y){
-                System.out.print(s + ", ");
-            }
-            System.out.print("\n");
-        }
-        ArrayList<String> z = x.genBoard();
-        assertNotSame(y, z);
     }
 }
