+++ app/src/test/java/com/example/swathi/gameboggle/TrieNodeTest.java
@@ -1,0 +1,14 @@
+package com.example.swathi.gameboggle;
+
+import org.junit.Before;
+
+import static org.junit.Assert.*;
+
+
+public class TrieNodeTest {
+    @Before
+    public void setUp() throws Exception {
+
+    }
+
+}
